nodes/dm-microphone/pyproject.toml
[project]
name = "dm-microphone"
version = "0.1.0"
description = "Microphone node with device selection and continuous audio streaming"
requires-python = ">=3.10"
dependencies = ["dora-rs >= 0.3.9", "pyarrow", "numpy", "sounddevice"]

[project.scripts]
dm-microphone = "dm_microphone.main:main"
